augmentatino 데이터 많이 늘린거로 학습했으나 시간이 너무 오래걸림

+++ features.py
@@ -79,12 +79,12 @@ def rotation(train, test):
 
 def terrain(train, test):
     a = (train[:, :, :, 9] / 100)
-    a = a.astype('int16')
+    a = a.astype('int8')
     a = a.reshape(75957, 40, 40, 1)
     train = np.concatenate((train, a), axis=-1)
 
     a = (test[:, :, :, 9] / 100)
-    a = a.astype('int16')
+    a = a.astype('int8')
     a = a.reshape(75957, 40, 40, 1)
     test = np.concatenate((test, a), axis=-1)
     del a
